download-and-convert-gdocs/requirements.txt
dill~=0.3.5.1
pip~=22.2.2
rsa~=4.9
pyasn1~=0.4.8
six~=1.16.0
cachetools~=5.2.0
certifi~=2022.9.14
urllib3~=1.26.12
requests~=2.28.1
google-api-python-client~=2.61.0
pyparsing~=3.0.9
httplib2~=0.20.4
oauthlib~=3.2.1
idna~=3.4
setuptools~=63.4.3
uritemplate~=4.1.1